Moved the test into the testtxn directory
